move common_utils test into folder
